useRef vs useState
useRef is used for updating the DOM without causing a re-render, while useState is used for rendering changes.
When using useRef to store a value when a button is clicked, the updated value will not appear in the UI directly. Instead, it can be displayed in the console or by using a button that triggers a re-render (e.g., with useState)